% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMPshiny.R
\name{DMPshiny}
\alias{DMPshiny}
\title{Launch DataMetProcess Shiny Application}
\usage{
DMPshiny(
  host = "127.0.0.1",
  port = NULL,
  launch.browser = TRUE,
  maxUploadSize = 200
)
}
\arguments{
\item{host}{Character. The host address where the application will run. Default is "127.0.0.1".}

\item{port}{Integer. The port on which the application will run. If NULL, a random port will be used.}

\item{launch.browser}{Logical. Indicates whether the browser should be launched automatically. Default is TRUE.}

\item{maxUploadSize}{Numeric. Maximum upload file size in megabytes. Default is 200.}
}
\value{
This function does not return a value. It starts the Shiny server and opens the application in the specified browser.
}
\description{
The `DMPshiny` function is used to start the Shiny application of the `DataMetProcess` package.
It allows configuring the host address, port, whether to launch the browser automatically,
and the maximum upload size.
}
\details{
The function sets Shiny options, such as the maximum upload size, and then runs the Shiny application located in the `DataMetProcess_Shiny/App.R` directory of the package.
}
\examples{
\dontrun{
  DMPshiny()
}

}
